+++ index.js
@@ -1,5 +1,7 @@
 exports.version = '0.14.0';
 
+exports.apiVersion = 'v2';
+exports.timeout = 2000;
 exports.cache = {};
 exports.user = {}
 exports.language = 'en-US'