+++ CHANGELOG.md
@@ -14,8 +14,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 ## [Unreleased]
 ### Fixed
-- language bug
-- parse functions bugs
+- language attribute
+- parse functions
 
 ## [0.12.0] - 2025-05-27
 ### Added