+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Fixed
+- language attribute
+- parse functions
 
 ## [0.14.0] - 2025-07-23
 ### Added