+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+## [0.15.0] - 2025-09-04
 ### Added
 - CreditSigner resendToken method
 ### Fixed